apps/api/pyproject.toml
[project]
name = "vitaltags-api"
version = "0.1.0"
description = "FastAPI backend for Vital Tags"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic-settings==2.4.0",
  "sqlmodel==0.0.22",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "itsdangerous==2.2.0",
  "redis==5.0.7",
  "rq==1.16.2",
  "minio==7.2.9",
  "reportlab==4.2.2",
  "qrcode[pil]==7.4.2",
  "orjson==3.10.7",
  "user-agents==2.2.0",
  "pycountry==24.6.1",
  "httpx==0.27.0",
  "python-multipart==0.0.9",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py311']